debounce the clockin button so that you can't double click
