+++ desktop/package.json
@@ -48,6 +48,8 @@
     "@types/node": "12.7.2",
     "@types/react": "^16.9.2",
     "@types/react-dom": "16.9.0",
+    "@types/react-redux": "^7.1.2",
+    "@types/react-router-dom": "^4.3.5",
     "axios": "^0.18.0",
     "classnames": "^2.2.6",
     "cross-env": "^5.2.0",
