--- desktop/src/constants/routes.js
@@ -1,9 +1,0 @@
-export const ROOT = `/`;
-
-export const EMPLOYEE = `employee`;
-export const ADMIN = `admin`;
-export const MANAGER = `manager`;
-
-export const CLOCKOUT = `clockout`;
-export const EXPORT = `export`;
-export const ANALYZE = `analyze`;
