+++ desktop/src/constants/routes.ts
@@ -1,0 +1,11 @@
+export const ROOT = `/`;
+
+export type AUTH_LEVELS = 'employee' | 'admin' | 'manager';
+
+export const EMPLOYEE = `employee`;
+export const ADMIN = `admin`;
+export const MANAGER = `manager`;
+
+export const CLOCKOUT = `clockout`;
+export const EXPORT = `export`;
+export const ANALYZE = `analyze`;
