--- desktop/src/routes/Admin/index.js
@@ -1,2 +1,0 @@
-import Admin from './admin';
-export default Admin;
