+++ desktop/src/routes/Admin/index.ts
@@ -1,0 +1,2 @@
+import Admin from './admin';
+export default Admin;
