--- desktop/src/routes/Employee/index.js
@@ -1,2 +1,0 @@
-import Employee from './employee';
-export default Employee;
