+++ desktop/src/routes/Employee/index.ts
@@ -1,0 +1,2 @@
+import Employee from './employee';
+export default Employee;
