--- desktop/src/routes/Manager/index.js
@@ -1,2 +1,0 @@
-import Manager from './manager';
-export default Manager;
