+++ desktop/src/routes/Manager/index.ts
@@ -1,0 +1,2 @@
+import Manager from './manager';
+export default Manager;
