--- desktop/src/scenes/Account/accountAction.js
@@ -1,16 +1,0 @@
-import React, { Component } from 'react';
-import PropTypes from 'prop-types';
-
-import AccountActionContainer from 'containers/Account/accountActionContainer';
-
-class AccountAction extends Component {
-  render() {
-    return <AccountActionContainer type={this.props.type} />;
-  }
-}
-
-AccountAction.propTypes = {
-  type: PropTypes.string.isRequired
-};
-
-export default AccountAction;
